pyproject.toml
[tool.poetry]
name = "iterm-smartopen"
version = "0.1.0"
description = ""
authors = ["Ulrich Petri <ulo@ulo.pe>"]

[tool.poetry.dependencies]
python = "^3.7"
py-applescript = "^1.0"
pyobjc = "^5.1.2"
click = "^7.0"
pasteboard = "^0.2.0"

[tool.poetry.dev-dependencies]
pytest = "^3.0"
bpython = "^0.17.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.poetry.scripts]
iterm-smartopen = "iterm_smartopen.main:main"
